fix(ffe-file-upload-react)!: make component props compliant with its type definitions
The `onFilesDropped` callback prop received a drop event as argument, not a `fileList` as described
in the type definitions. Typescript clients would have to circumvent the mismatch by casting to
the `any` type.

BREAKING CHANGE: The `onFilesDropped` callback prop is now called with a `fileList` instead
of the drop event