feat(ffe-account-selector-react): add displayAttribute prop
makes it possible to override what is displayed in the AccountSelector input.
So you can have a new custom attribute in the input, example a new attribute that contains: account_name - account_number

previous default was the first attribute in the searchAttribute prop. The same should be the case with this change.

Needed to add the displayAttribute to the searchAttributes passed to the SearchableDropdown, so the user don't see a message about 0 hits when changing the input.